Conflicts:
	src/com/bsb/hike/chatthread/OfflineChatThread.java

+++ src/com/bsb/hike/offline/HikeConverter.java
@@ -2,12 +2,10 @@ package com.bsb.hike.offline;
 
 import java.io.File;
 
-import org.eclipse.paho.client.mqttv3.MqttMessage;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import android.content.Context;
-import android.graphics.drawable.GradientDrawable.Orientation;
 import android.util.Pair;
 
 import com.bsb.hike.HikeConstants;
@@ -17,9 +15,6 @@ import com.bsb.hike.db.HikeConversationsDatabase;
 import com.bsb.hike.filetransfer.FileTransferManager;
 import com.bsb.hike.models.ConvMessage;
 import com.bsb.hike.models.ConvMessage.OriginType;
-import com.bsb.hike.models.ConvMessage.State;
-import com.bsb.hike.models.HikeFile;
-
 import com.bsb.hike.models.HikeFile.HikeFileType;
 import com.bsb.hike.service.MqttMessagesManager;
 import com.bsb.hike.utils.Logger;
@@ -76,7 +71,7 @@ public class HikeConverter implements IMessageReceived, IMessageSent {
 				convMessage.serialize().toString(), OfflineConstants.FILE_TOPIC)
 				.file(file).persistance(false).ackRequired(true).build();
 		senderConsignment.setTag(convMessage);
-		senderConsignment.setAwb((int) convMessage.getMsgID());
+		senderConsignment.setAwb(convMessage.getMsgID());
 		
 		FileTransferModel fileTransferModel = new FileTransferModel(
 				new TransferProgress(0, OfflineUtils.getTotalChunks((int) file
@@ -312,7 +307,7 @@ public class HikeConverter implements IMessageReceived, IMessageSent {
 					messageJSON.toString(), OfflineConstants.TEXT_TOPIC)
 					.build();
 		}
-		senderConsignment.setAwb((int) convMessage.getMsgID());
+		senderConsignment.setAwb(convMessage.getMsgID());
 		senderConsignment.setTag(convMessage);
 		return senderConsignment;
 	}
